+++ src/backend/marsha/core/serializers/__init__.py
@@ -6,6 +6,7 @@ from .file import *  # noqa isort:skip
 from .live_session import *  # noqa isort:skip
 from .pairing_challenge import *  # noqa isort:skip
 from .playlist import *  # noqa isort:skip
+from .playlist_access import *  # noqa isort:skip
 from .portability_request import *  # noqa isort:skip
 from .shared_live_media import *  # noqa isort:skip
 from .thumbnail import *  # noqa isort:skip